+++ assets/javascripts/kitten/index.js
@@ -111,6 +111,8 @@ export { Grid, GridCol } from './components/grid/grid'
 export { LegoGrid } from './components/grid/lego-grid'
 export { Row } from './components/grid/row'
 export { SideGrid } from './components/grid/side-grid'
+export { StickyContainer } from './components/grid/sticky-container'
+export { AdaptableGrid } from './components/grid/adaptable-grid'
 
 // Headers
 export { HeaderLogo } from './components/headers/header-logo'