+++ src/index.js
@@ -477,6 +477,18 @@ Object.defineProperty(exports, "SideGrid", {
     return _sideGrid.SideGrid;
   }
 });
+Object.defineProperty(exports, "StickyContainer", {
+  enumerable: true,
+  get: function get() {
+    return _stickyContainer.StickyContainer;
+  }
+});
+Object.defineProperty(exports, "AdaptableGrid", {
+  enumerable: true,
+  get: function get() {
+    return _adaptableGrid.AdaptableGrid;
+  }
+});
 Object.defineProperty(exports, "HeaderLogo", {
   enumerable: true,
   get: function get() {
@@ -1021,6 +1033,10 @@ var _legoGrid = require("./components/grid/lego-grid");
 var _row = require("./components/grid/row");
 
 var _sideGrid = require("./components/grid/side-grid");
+
+var _stickyContainer = require("./components/grid/sticky-container");
+
+var _adaptableGrid = require("./components/grid/adaptable-grid");
 
 var _headerLogo = require("./components/headers/header-logo");