+++ CHANGELOG.md
@@ -5,7 +5,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add player button on `Video` component.
-- Fix: Remove `getScssPaths` from `@kisskissbankbank/kitten`.
+
+## [2.1.0] - 2019-04-11
+
+Features:
+- New `StickyContainer` component.
+- Add a new component `AdaptableGrid` to handle grids.
+  From now, you can use this component instead of `Grid`.
+
+## [2.0.1] - 2019-04-11
+
+Fix:
+- Remove `getScssPaths` from `@kisskissbankbank/kitten`.
   Prefer use `@kisskissbankbank/kitten/src/config/paths`.
 
 ## [2.0.0] - 2019-04-10